requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
beautifulsoup4==4.12.3
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
playwright==1.49.1
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyee==12.0.0
python-dotenv==1.0.1
setuptools==75.6.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.45.1
